feat(model): :sparkles: add variable to choose the number of each agents types
